fix: metadata bool variables should not also impment display logic
There may be some cases where we don't want to display a custom
copyright statement even when one is set, but we shouldn't package
that logic into a bool value called $has_custom_copyright because
it pollutes the facial meaning of the variable name and hides the
display logic.

In the cases where we don't want to display a custom copyright that
has been set, consult the $show_custom_copyright variable.